+++ src/mpid/ch4/src/ch4r_init.h
@@ -15,6 +15,8 @@
 #include "ch4i_util.h"
 #include "ch4r_buf.h"
 #include "ch4r_callbacks.h"
+#include "ch4r_rma_target_callbacks.h"
+#include "ch4r_rma_origin_callbacks.h"
 #include "mpl_uthash.h"
 
 #undef FUNCNAME
